+++ server/src/main/java/com/google/codeu/controllers/blobstore/BlobstoreHelper.java
@@ -52,67 +52,6 @@ public class BlobstoreHelper {
      * @param imageStreams List of byte streams describing the input images.
      * @return An array of <code>URL</code> to the location of the uploaded images.
      */
-    /**@Override
-    public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        PrintWriter out = response.getWriter();
-        //Get message entered by the user
-        //String message = request.getParameter("message");
-        //Get Blobkey that points to the image uploaded by user
-        BlobKey blobKey = getBlobKey(request, "image");
-        //User did not upload a file, so render an error msg
-        if (blobKey == null) {
-            //out.println("Please upload an image file");
-            response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
-            return;
-        }
-        //Get the URL of image that the user uploaded
-        String imageUrl = getUploadedFileUrl(blobKey);
-        //Get the labels of the image that the user uploaded
-        byte[] imageStreams = getBlobBytes(blobKey);
-        List<EntityAnnotation> imageLabels = getImageLabels(imageStreams);
-        
-        //Output some HTML that shows the data that user entered
-        
-        response.setContentType("text/html");
-        out.println("<p>Here's the image that you uploaded:</p>");
-        out.println("<a href=\"" + imageUrl + "\">");
-        out.println("<img src=\"" + imageUrl + "\" />");
-        out.println("</a>");
-        out.println("<p>Here are the labels we extracted:</p>");
-        out.println("<ul>");
-        for(EntityAnnotation label : imageLabels){
-            out.println("<li>" + label.getDescription() + " " + label.getScore());
-        }
-        out.println("</ul>");
-    }*/
-
-    private BlobKey getBlobKey(HttpServletRequest request, String formInputElementName) {
-        BlobstoreService blobstoreService = BlobstoreServiceFactory.getBlobstoreService();
-        Map<String, List<BlobKey>> blobs = blobstoreService.getUploads(request);
-        List<BlobKey> blobKeys = blobs.get("image");
-
-        /**
-         * User submitted form without selecting a file
-         * so we cannot get a blobKey (devserver)
-         */
-        if(blobKeys == null || blobKeys.isEmpty()) {
-            return null;
-        }
-
-        // Our form only contains a single file input, get first index
-        BlobKey blobKey = blobKeys.get(0);
-
-        /**
-         * User submitted a form without selecting a file
-         * so the BlobKey is empty (live server)
-         */
-        BlobInfo blobInfo = new BlobInfoFactory().loadBlobInfo(blobKey);
-        if (blobInfo.getSize() == 0) {
-            blobstoreService.delete(blobKey);
-            return null;
-        }
-        return blobKey;
-    }
 
     private byte[] getBlobBytes(BlobKey blobKey) throws IOException {
         BlobstoreService blobstoreService = BlobstoreServiceFactory.getBlobstoreService();
