+++ server/src/main/webapp/image-processing.jsp
@@ -1,0 +1,22 @@
+<%@ page import="com.google.appengine.api.blobstore.BlobstoreService" %>
+<%@ page import="com.google.appengine.api.blobstore.BlobstoreServiceFactory" %>
+<% BlobstoreService blobstoreService = BlobstoreServiceFactory.getBlobstoreService();
+   String uploadUrl = blobstoreService.createUploadUrl("/image-processing"); %>
+
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="UTF-8">
+    <title>Image Upload Analysis</title>
+  </head>
+  <body>
+    <h1>Image Upload Analysis</h1>
+
+    <form method="POST" enctype="multipart/form-data" action="<%= uploadUrl %>">
+      <p>Upload an image:</p>
+      <input type="file" name="image">
+      <br/><br/>
+      <button>Submit</button>
+    </form>
+  </body>
+</html>
