+++ server/src/main/webapp/index.html
@@ -34,6 +34,7 @@ limitations under the License.
         <li><a href="/geomessage-feed.html">Map - Part 3 Demo</a></li>
         <li><a href="/chart.html">Chart - Part 2 Demo</a></li>
         <li><a href="/fetch-blobstore.html">Images - Part 2 Demo</a></li>
+        <li><a href="/image-processing.jsp">Images Testing</a><li>
       </ul>
     </nav>
     <h1>CodeU Starter Project</h1>
